requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
asn1crypto==0.24.0
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
cryptography==2.1.4
flake8==3.8.4
flake8-polyfill==1.0.2
future==0.18.2
gast==0.3.3
git-python==1.0.3
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.6
idna-ssl==1.1.0
importlib-metadata==2.0.0
iniconfig==1.1.1
invoke==1.4.1
isort==5.6.4
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.4.3
mando==0.6.4
Markdown==3.2.2
mccabe==0.6.1
multidict==4.7.6
mypy==0.790
mypy-extensions==0.4.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycrypto==2.6.1
pyflakes==2.2.0
pygobject==3.26.1
pylint==2.6.0
pyparsing==2.4.7
pytest==6.1.2
python-apt==1.6.5+ubuntu0.3
pyxdg==0.25
PyYAML==5.3.1
radon==4.3.2
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
SecretStorage==2.3.1
six==1.15.0
smmap==3.0.4
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.1
termcolor==1.1.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
xenon==0.7.1
yarl==1.6.0
zipp==3.2.0
